requirements.txt
absl-py==1.4.0
addict==2.4.0
black==22.8.0
blessed==1.20.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
coloredlogs==15.0.1
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.8
easydict==1.9
filelock==3.4.1
flake8==5.0.4
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.16.1
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.48.2
huepy==1.2.1
huggingface-hub==0.4.0
humanfriendly==10.0
idna==3.4
importlib-metadata==4.2.0
importlib-resources==5.4.0
iopath==0.1.10
isort==5.10.1
joblib==1.1.1
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mccabe==0.7.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.7.1
model-index==0.1.11
more-itertools==8.12.0
mypy-extensions==1.0.0
numpy==1.19.5
nvidia-ml-py==11.495.46
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
openmim==0.3.6
ordered-set==4.0.2
packaging==21.3
pandas==1.1.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
portalocker==2.7.0
protobuf==3.19.6
psutil==5.9.4
ptflops==0.6.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.14.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
timm==0.6.13
tomli==1.2.3
torch==1.8.2
torchaudio==0.8.2
torchprofile==0.0.4
torchstat==0.0.7
torchvision==0.9.2
tornado==6.1
tqdm==4.64.1
typed-ast==1.5.4
typing-extensions==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.0.3
yacs==0.1.8
yapf==0.32.0
zipp==3.6.0
